requirements.txt
coreschema==0.0.4
Django==4.2.5
django-filter==23.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
openapi-codec==1.3.2
psycopg==3.1.12
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
simplejson==3.19.1
sqlparse==0.4.4
Pillow